Moved log.go into its own module. I suspect we'll want more things in here
